requirements.txt
pandas
numpy>=1.24.0
pandas-ta
requests
python-dotenv
flask
flask-cors
flask-limiter
pyopenssl
gunicorn
flask-login
google-auth
google-auth-oauthlib
google-auth-httplib2
flask-sqlalchemy
psutil
